💫add bubble card & modify components name to avoid misusing & optimize images shown in the card
