+++ src/components/bubble-card/bubble-answer/BubbleAnswer.css
@@ -27,10 +27,6 @@
   }
 }
 
-.card-body {
-  padding: "14px 18px 0px 18px";
-}
-
 .ans-title {
   font-size: 16px;
   font-weight: bold;
@@ -39,7 +35,11 @@
 }
 
 .ans-body {
-  max-width: 100%;
+  /* background-color: #77b9ff; */
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
 
   .quote-container {
     background-color: #add3fe69;
@@ -80,11 +80,18 @@
   }
 
   .photos-container {
+    width: 100%;
     display: flex;
-    flex-wrap: wrap;
-    margin-top: 10px;
+    justify-content: center;
     margin-bottom: 10px;
-    gap: 10px 10px;
+
+    .preview-group {
+      /* background-color: #333; */
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: flex-start;
+      gap: 10px 10px;
+    }
   }
 }
