+++ src/components/bubble-card/bubble-basic-card/BubbleCard.css
@@ -13,6 +13,40 @@
   transform: translateY(-4px);
 }
 
+.card-body {
+  position: relative;
+  padding: 14px 20px 14px 20px;
+
+  .news-dot {
+    position: absolute;
+    top: 18px;
+    left: 6px;
+    width: 10px;
+    height: 10px;
+    border-radius: 50%;
+    background-color: #66b0ff;
+  }
+
+  .pin {
+    position: absolute;
+    top: 6px;
+    right: 6px;
+    cursor: pointer;
+  }
+
+  .pin:hover::after {
+    content: attr(data-tips);
+    position: absolute;
+    font-size: 12px;
+    white-space: nowrap;
+    top: -30%;
+    left: 110%;
+    padding: 2px;
+    border: #66b0ff solid 1px;
+    color: #66b0ff;
+  }
+}
+
 .q-title {
   font-size: 16px;
   font-weight: bold;
@@ -21,6 +55,9 @@
 }
 
 .q-body {
+  display: flex;
+  flex-direction: column;
+
   .text {
     white-space: pre-wrap;
     display: block;
@@ -32,10 +69,15 @@
 
   .photos-container {
     display: flex;
-    flex-wrap: wrap;
-    margin-top: 10px;
+    justify-content: center;
     margin-bottom: 10px;
-    gap: 10px 10px;
+
+    .preview-group {
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: flex-start;
+      gap: 10px 10px;
+    }
   }
 }
 
@@ -60,53 +102,5 @@
     font-size: 12px;
     color: #818181;
     /* margin-right: 16px; */
-  }
-}
-
-.card-footer {
-  width: 100%;
-  height: 42px;
-  border-top: #eeeded solid 1px;
-  display: flex;
-  align-items: center;
-  margin-top: 5px;
-  padding-bottom: 5px;
-
-  .avatar {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    box-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
-  }
-
-  .msg-icon {
-    margin-left: 16px;
-  }
-
-  .text {
-    margin-left: 12px;
-    font-size: 16px;
-    color: #66b0ff;
-  }
-
-  .favourite {
-    position: relative;
-    margin-left: auto;
-    margin-right: 16px;
-    margin-top: 5px;
-    cursor: pointer;
-  }
-
-  .favourite:hover::after {
-    content: attr(data-tips);
-    position: absolute;
-    font-size: 12px;
-    white-space: nowrap;
-    top: -50%;
-    left: 100%;
-    padding: 2px;
-    background-color: #fff;
-    border: #66b0ff solid 1px;
-    color: #66b0ff;
   }
 }
